DrinkReviews functional is ready

+++ GetDrinksService/Requests/UpdateDrinkRequest.cs
@@ -6,7 +6,7 @@ namespace DrinksService
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Strength { get; set; }
+        public double Strength { get; set; }
         public ProjectEnums.DrinkCategory Category { get; set; }
         public string Description { get; set; }
     }
